[EXPIRMENTAL BRANCH] Split OpenCL and OpenMP modules up.  Next steps: implement MMulGPU an MMulOMP and SolverFactory per doc, test on AWS GPU Cluster.
