+++ pom.xml
@@ -853,6 +853,7 @@
     <!--<module>flink</module>-->
     <!--<module>h2o</module>-->
     <module>viennacl</module>
+    <module>viennacl-omp</module>
   </modules>
 
   <profiles>
