+++ viennacl-omp/runs
@@ -1,0 +1,32 @@
+original 
+row-major viennacl::matrix
+  + OCL matrix memory domain after assgn=2 
+- dense vcl mmul with fast_copy
+- mmul microbenchmark
+  +  Mahout multiplication time: 15699 ms. 
+  + ViennaCL/OpenCL multiplication time: 3625 ms. 
+  + ompA mem domain:1 
+  + ompB mem domain:1 
+  + ViennaCL/cpu/OpenMP multiplication time: 2838 ms. 
+
+with sys.ArrayCopy, all dense.
+ViennaCLSuite:
+- row-major viennacl::matrix
+  + OCL matrix memory domain after assgn=2 
+- dense vcl mmul with fast_copy
+- mmul microbenchmark
+  +  Mahout multiplication time: 15407 ms. 
+  + ViennaCL/OpenCL multiplication time: 3499 ms. 
+  + ompA mem domain:1 
+  + ompB mem domain:1 
+  + ViennaCL/cpu/OpenMP multiplication time: 2714 ms.
+
+DL latest
+ViennaCLSuite:
+- row-major viennacl::matrix
+  + OCL matrix memory domain after assgn=2 
+- dense vcl mmul with fast_copy
+- mmul microbenchmark
+  +  Mahout multiplication time: 16076 ms. 
+  + ViennaCL/OpenCL multiplication time: 3360 ms. 
+  + ViennaCL/cpu/OpenMP multiplication time: 2666 ms.
