--- viennacl/src/main/java/org/apache/mahout/viennacl/javacpp/MatrixTransExpression.scala
@@ -1,18 +1,0 @@
-package org.apache.mahout.viennacl.javacpp
-
-import org.bytedeco.javacpp.Pointer
-import org.bytedeco.javacpp.annotation.{Name, Namespace, Platform, Properties}
-
-
-@Properties(inherit = Array(classOf[Context]),
-  value = Array(new Platform(
-    include = Array("matrix.hpp"),
-    library = "jniViennaCL")
-  ))
-@Namespace("viennacl")
-@Name(Array("matrix_expression<const viennacl::matrix_base<double>, " +
-  "const viennacl::matrix_base<double>, " +
-  "viennacl::op_trans>"))
-class MatrixTransExpression extends Pointer {
-
-}
