--- viennacl/src/main/scala/org/apache/mahout/viennacl/javacpp/MemHandle.scala
@@ -1,18 +1,0 @@
-package org.apache.mahout.viennacl.javacpp
-
-import org.bytedeco.javacpp.Pointer
-import org.bytedeco.javacpp.annotation._
-
-
-@Properties(inherit = Array(classOf[Context]),
-  value = Array(new Platform(
-    library = "jniViennaCL")
-  ))
-@Namespace("viennacl::backend")
-@Name(Array("mem_handle"))
-class MemHandle extends Pointer {
-
-  allocate()
-
-  @native def allocate()
-}
