--- viennacl/src/main/scala/org/apache/mahout/viennacl/javacpp/VecMultExpression.scala
@@ -1,16 +1,0 @@
-package org.apache.mahout.viennacl.javacpp
-
-import org.bytedeco.javacpp.Pointer
-import org.bytedeco.javacpp.annotation.{Name, Namespace, Platform, Properties}
-
-
-@Properties(inherit = Array(classOf[Context]),
-  value = Array(new Platform(
-    library = "jniViennaCL")
-  ))
-@Namespace("viennacl")
-@Name(Array("vector_expression<const viennacl::vector_base<double>," +
-  "const double, viennacl::op_mult >"))
-class VecMultExpression extends Pointer {
-
-}
